refactor(variant): reorganize files for reviewability
Two layout fixes after the incremental build-up:

- variant_column_copy.go had grown three concerns: the CopyVariantRows
  copier, the Writer.WriteRowGroup fast-path integration, and (after the
  integration section) the typed emitter the copier uses. The Writer
  integration moves to writer_variant.go, following the convention that
  <upstream file>_variant.go extends an upstream file; the copy file now
  reads top to bottom as one subsystem. The internal test moves with it
  (writer_variant_internal_test.go).
- Test helpers used across files (the randomized generators, raw variant
  row shapes, file builders, the cursor read-back oracle, and the
  assertions built on it) had accreted in whichever test file needed
  them first, forcing readers of merge_variant_test.go to chase helpers
  through four other files. They now live in variant_fixtures_test.go;
  helpers used by a single file stay in that file.
