+++ writer_variant_internal_test.go
